Move helper functions to helper.ts
