+++ lib/CotterVerify.js
@@ -13,6 +13,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const enum_1 = __importDefault(require("./enum"));
+const helper_1 = require("./helper");
 class Cotter {
     constructor(config) {
         console.log("Using origin: ", new URL(window.location.href).origin);
@@ -32,7 +33,7 @@ class Cotter {
             localStorage.setItem("COTTER_STATE", this.state);
         }
         // SUPPORT PKCE
-        this.verifier = generateVerifier();
+        this.verifier = helper_1.generateVerifier();
         this.loaded = false;
         this.cotterIframeID =
             Math.random().toString(36).substring(2, 15) + "cotter-signup-iframe";
@@ -150,7 +151,7 @@ class Cotter {
             ifrm.style.minHeight = "520px";
         }
         ifrm.style.overflow = "scroll";
-        challengeFromVerifier(this.verifier).then((challenge) => {
+        helper_1.challengeFromVerifier(this.verifier).then((challenge) => {
             ifrm.setAttribute("src", encodeURI(`${enum_1.default.CotterBaseURL}/signup?challenge=${challenge}&type=${this.config.Type}&domain=${this.config.Domain}&api_key=${this.config.ApiKeyID}&redirect_url=${this.config.RedirectURL}&state=${this.state}&id=${this.config.ContainerID}`));
         });
         ifrm.setAttribute("allowtransparency", "true");
@@ -283,34 +284,4 @@ class Cotter {
     }
 }
 const isIFrame = (input) => input !== null && input.tagName === "IFRAME";
-function dec2hex(dec) {
-    return ("0" + dec.toString(16)).substr(-2);
-}
-function generateVerifier() {
-    var array = new Uint32Array(56 / 2);
-    window.crypto.getRandomValues(array);
-    return Array.from(array, dec2hex).join("");
-}
-function sha256(plain) {
-    // returns promise ArrayBuffer
-    const encoder = new TextEncoder();
-    const data = encoder.encode(plain);
-    return window.crypto.subtle.digest("SHA-256", data);
-}
-function base64urlencode(a) {
-    var str = "";
-    var bytes = new Uint8Array(a);
-    var len = bytes.byteLength;
-    for (var i = 0; i < len; i++) {
-        str += String.fromCharCode(bytes[i]);
-    }
-    return btoa(str).replace(/\+/g, "-").replace(/\//g, "_").replace(/=+$/, "");
-}
-function challengeFromVerifier(v) {
-    return __awaiter(this, void 0, void 0, function* () {
-        var hashed = yield sha256(v);
-        var base64encoded = base64urlencode(hashed);
-        return base64encoded;
-    });
-}
 exports.default = Cotter;
